requirements.txt
certifi==2021.5.30
chardet==3.0.4
idna==2.8
numpy==1.20.2
opencv-python==4.5.2.54
praw==7.3.0
prawcore==2.2.0
PySimpleGUI==4.45.0
python-dotenv==0.18.0
requests==2.22.0
requests-unixsocket==0.2.0
update-checker==0.18.0
urllib3==1.25.11
websocket-client==1.1.0
